requirements.txt
-f /usr/share/pip-wheels
asgiref==3.3.1
Django==3.1.3
django-generate-secret-key==1.0.2
httplib2==0.18.1
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
